Revert "[add, fix] 버그 수정 및 관련 위키리스트 목록 "
