+++ src/components/main/WikiRow.tsx
@@ -3,14 +3,15 @@ import { Wiki } from "../../interfaces/Data";
 import { useNavigate } from "react-router-dom";
 
 interface Props {
+  wikiList: Wiki[];
   item: Wiki;
 }
 
-const WikiRow = ({ item }: Props) => {
+const WikiRow = ({ wikiList, item }: Props) => {
   const navigate = useNavigate();
 
   const onClick = () => {
-    navigate(`/wiki/${item.id}`);
+    navigate(`/wiki/${item.id}`, { state: { wikiList: wikiList } });
   };
 
   return (
