+++ src/pages/Main.tsx
@@ -4,11 +4,31 @@ import { Wiki } from "../interfaces/Data";
 import { getAllWiki, postWiki } from "../api/wiki";
 import WikiList from "../components/main/WikiList";
 import WikiListToolbar from "../components/main/WikiListToolbar";
-import useWikiList from "../components/hooks/useWikiList";
 
 const Main = () => {
-  const { wikiList, loading, pending, fetchWikiList, createWiki } =
-    useWikiList();
+  const [wikiList, setWikiList] = useState<Wiki[] | null>(null);
+  const [loading, setLoading] = useState<boolean>(false);
+  const [pending, setPending] = useState<boolean>(false);
+
+  const fetchWikiList = async () => {
+    setLoading(true);
+
+    const wikiListRes = await getAllWiki();
+    if (wikiListRes) {
+      setWikiList(wikiListRes);
+      setLoading(false);
+    }
+  };
+
+  const createWiki = async (arg: Wiki) => {
+    const resultCreateWiki = await postWiki(arg);
+
+    setPending(true);
+    if (resultCreateWiki === "success") {
+      setPending(false);
+      fetchWikiList();
+    }
+  };
 
   useEffect(() => {
     fetchWikiList();
